backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv

notion-client

redis==5.0.4